Use constexpr to check if experimental features are enabled

+++ test/src/unit-backwards_compat.cc
@@ -684,50 +684,50 @@ TEST_CASE(
 TEST_CASE(
     "Backwards compatibility: Write to an array of older version",
     "[backwards-compat][write-to-older-version]") {
-#ifndef TILEDB_EXPERIMENTAL_FEATURES
-  std::string old_array_name(arrays_dir + "/non_split_coords_v1_4_0");
-  Context ctx;
-  std::string fragment_uri;
+  if constexpr (!is_experimental_build) {
+    std::string old_array_name(arrays_dir + "/non_split_coords_v1_4_0");
+    Context ctx;
+    std::string fragment_uri;
 
-  try {
-    // Write
-    Array old_array(ctx, old_array_name, TILEDB_WRITE);
-    std::vector<int> a_write = {100};
-    std::vector<int> coords_write = {1, 10};
-    Query query_w(ctx, old_array);
-    query_w.set_layout(TILEDB_UNORDERED)
-        .set_data_buffer("a", a_write)
-        .set_coordinates(coords_write);
-    query_w.submit();
-    old_array.close();
+    try {
+      // Write
+      Array old_array(ctx, old_array_name, TILEDB_WRITE);
+      std::vector<int> a_write = {100};
+      std::vector<int> coords_write = {1, 10};
+      Query query_w(ctx, old_array);
+      query_w.set_layout(TILEDB_UNORDERED)
+          .set_data_buffer("a", a_write)
+          .set_coordinates(coords_write);
+      query_w.submit();
+      old_array.close();
 
-    // Read
-    std::vector<int> subarray = {1, 4, 10, 10};
-    std::vector<int> a_read(50);
-    std::vector<int> coords_read(50);
+      // Read
+      std::vector<int> subarray = {1, 4, 10, 10};
+      std::vector<int> a_read(50);
+      std::vector<int> coords_read(50);
 
-    Array array(ctx, old_array_name, TILEDB_READ);
-    Query query_r(ctx, array);
-    query_r.set_subarray(subarray)
-        .set_layout(TILEDB_ROW_MAJOR)
-        .set_data_buffer("a", a_read)
-        .set_coordinates(coords_read);
-    query_r.submit();
-    array.close();
+      Array array(ctx, old_array_name, TILEDB_READ);
+      Query query_r(ctx, array);
+      query_r.set_subarray(subarray)
+          .set_layout(TILEDB_ROW_MAJOR)
+          .set_data_buffer("a", a_read)
+          .set_coordinates(coords_read);
+      query_r.submit();
+      array.close();
 
-    // Remove created fragment and ok file
-    VFS vfs(ctx);
-    vfs.remove_dir(get_fragment_dir(old_array_name));
-    vfs.remove_dir(get_commit_dir(old_array_name));
+      // Remove created fragment and ok file
+      VFS vfs(ctx);
+      vfs.remove_dir(get_fragment_dir(old_array_name));
+      vfs.remove_dir(get_commit_dir(old_array_name));
 
-    REQUIRE(a_read[0] == 100);
-    for (int i = 1; i < 4; i++) {
-      REQUIRE(a_read[i] == i + 1);
+      REQUIRE(a_read[0] == 100);
+      for (int i = 1; i < 4; i++) {
+        REQUIRE(a_read[i] == i + 1);
+      }
+    } catch (const std::exception& e) {
+      CHECK(false);
     }
-  } catch (const std::exception& e) {
-    CHECK(false);
-  }
-#endif
+  }  // !is_experimental_build
 }
 
 TEST_CASE(
